(PE-2790) Fix for Jetty always opening binding for a plain http port
Prior to this commit, Jetty would always open a binding for an http
port, whether or not a "port" value was actually specified in the
webserver config for the service.

With the changes in this commit, an http port binding is only created if
one is specified in the webserver config.