(PE-2866) Prepare trapperkeeper-webserver-jetty9 for 0.3.3 release
Updated CHANGELOG.

Added lein-project-version and lein-release plugin references to the
project.clj file.

Prior to this commit, this package depended upon puppetlabs/kitchensink
0.5.1.  This commit updates the dependency version to 0.5.2.

Prior to this commit, the jenkins deploy.sh script was written to
try to commit package versioning changes on the Jenkins build machine
based on the current branch.  In this commit, the jenkins deploy.sh
script was updated to use HEAD:master instead since Jenkins runs against
a detached HEAD instead of a branch.