(PE-2650) bootstrap-services-with-cli-data -> with-app-with-cli-data
This commit updates the tests to use the safer with-app-with-cli-data
function for booting a TK service to be tested instead of the older
bootstrap-services-with-cli-data.
As the new function does a better job at setting up/tearing down the test
environment it was also possible to simplify the tests and while at it
some unused imports/requires were removed.