Reorganize service tests
This commit doesn't contain any code changes; it just reorganizes
the tests a bit so that you can run them with a little more
granularity.  In particular, it separates all of the SSL tests
that are intended to result in a successful HTTPS request into
one deftest, and all of the ones that expect a failed HTTPS
request into a separate deftest.