(PE-3401) (REVIEW ONLY) Expose webserver setting override function
This commit includes a new function on the webserver service,
`override-webserver-settings!`, which can be used to override any of the
settings in the webserver section of the service's Trapperkeeper config
file prior to the webserver being started.  This commit has a proposed
implementation and tests but, assuming it is workable, would need to be
supplemented with some documentation in the README.md on how use the new
function.

This commit also removes the `certificate-whitelist` setting from the
jetty-config.md file as it was found to not be used anywhere in the
webserver implementation.