(TK-2) Rename `:config` atom to `:state`
We need a place to store the original configuration data in the
context map, and the logical name for that is `:config`; we were
using that for another purpose (to store an atom that had some
mutable state).  This commit simply changes the existing uses
to use the key `:state` instead, which is probably a bit more
accurate of a description of what we're doing, and also frees
us up to be able to use `:config` to store the actual config
in subsequent commits.