(PE-3140) Wrap webserver start in shutdown-on-error
This commit wraps the `start-webserver` call during the `start` phase of
the webserver-jetty9 service in a `shutdown-on-error` call.  This allows
the Jetty server instance to be persisted to the service context and,
therefore, have a `stop` call for the service stop the Jetty server
instance.  Stopping of the Jetty server instance is necessary in order
for the JVM process in which the server was started to gracefully exit.