(PE-3659) Added Jetty server validation to service `init` exception test
This commit adds an assertion to a test that no Jetty server instance was
attached to the service context when a dependent service throws an
exception during its `init` function.

The commit also adjusts the test description to validate what it is
actually doing around the validation of the Jetty server instance.