(PE-3140) Catch `start-webserver` errors in `start`, calling `shutdown`
This commit changes the implementation of error handling for the call to
`start-webserver` during the start phase of the service to call
`shutdown` explicitly.  The implementation in the previous commit relied
upon the Trapperkeeper framework to invoke `shutdown` implicitly when
the service was being shut down -- signalled through the exception being
caught in a `shutdown-on-error` block.