(PE-3914) Specify crl-path from Jetty config
This commit enables a new setting in the webserver config, crl-path,
which can be set in order for Jetty's SSL server binding to load and use
the CRLs from the specified file to validate incoming client
connections.

This commit also broadens the class caught on Jetty startup failures to
a full Exception.  In the previous commit, the class was just an
IOException.  Jetty can throw CRLException (or possibly other
exceptions) on processing a CRL file during initialization, so the
broader class being caught allows for the service to be cleaned up
properly in response to a failure during startup to process CRLs.