Make request header size configurable
Add a new configuration option, :request-size, to the webserver
configuration. This will allow the user to increase the size
of the request header when they are sending an extremely
large cookie as part of a request, as if this is too
small, an HTTP 413 error will be returned. This should not need
to be increased except in cases with unusually large headers.