Remove configurations from common.clj
Remove various configurations from common.clj and move them
into the files that are using them. The greater specificity
required when specifying a service in the configuration file
means that these configurations are now specific to particular
namespaces and are only being used in one place. Hence, they
are no longer common.