(PE-5503) Support for configuring a server certificate chain
This commit adds support for configuring a server certificate chain
(including the end certificate and any CA certificates) which the
server can deliver to the client for the client's use in validating the
server.

Previously, only a single server certificate could be configured,
without the full chain of trust.  The commit adds support for a new
optional `ssl-cert-chain` setting which can be used to configure the
set of CA certificates to append to the certificate from the `ssl-cert`
setting.