(TK-84) Don't double encode query strings
The Java URI class encodes the "query" parameter when using its
multi-parameter constructor. In proxy-servlet we are using getQueryString
to get the query string to pass into URI, which does not URL decode the
query string. This change decodes the query string before passing it into
the URI class, which will then go ahead and reencode it.