(TK-96) Fixes for default SSL protocols
This commit changes our default SSL protocols to include only
TLSv1, TLSv1.1, and TLSv1.2.  It also ensures that the server
is initialized with these protocols if the user doesn't specify
their own list via configuration.

Lastly, it makes sure that we use that same list of default
protocols for `add-proxy-route`, in the case where the user
has specified an explicit list of pem files for SSL but has
not specified protocols.