(TK-119) Remove some sleeps from the graceful shutdown tests
This commit replaces a sleep, done inside of the graceful shutdown tests
to ensure that the request is in process at the point the server
shutdown is initiated, with the use of a Clojure promise.  The Clojure
promise would better ensure that the Ring handler on the server had been
engaged at the point the server shutdown is triggered.