(TK-144) Add support for configuring Jetty via script
A problem that we have had a few times with Jetty is that, in order to try to
make it easy for our users to manage the application without having to be Jetty
experts, we run Jetty embedded and only surface a subset of its configuration
options that we believe the users might care about.

This provides a pretty nice UX in cases where we've plumbed the appropriate
settings through into our config before shipping a build, but, it can be
problematic when we are trying to debug a customer issue and we discover a
potentially-relevant Jetty setting that we haven't wired through to our
configuration files... it means that we can't toggle the Jetty config setting
without shipping them a new build.

After discovering how logback supports some basic Java scripting in their
EvaluatorFilters (via a library called Janino), it occurred to me that we might
be able to support something similar for our Jetty configuration.

This commit does that.  It creates a setting called 'post-config-script', which
can be used for advanced use cases to provide some code to manipulate a Jetty
`Server` object programatically. This should theoretically allow us to cobble
together some sample configuration to ship to users who already have a build
installed that can modify basically any setting that is available in Jetty.

Use cases for this setting would hopefully be temporary, and when we find new
Jetty settings that we determine would be valuable to users, we can just plumb
them through to our normal configuration for the next release, but this would
give users a workaround that doesn't require us to ship them a whole new build
in the interim.