(MAINT) Update deps and options for running SSLv3 unit tests
This commit bumps the dependencies in the project.clj file.  clj-time is
specifically pinned to 0.5.1 for compatibility with other current
Trapperkeeper library dependencies and to avoid the conflict that was
being pulled in from spyscope's dependency on clj-time.  The clojure
dependency was bumped to 1.6.0 since other Trapperkeeper projects are
incrementally having this bumped as well.

This commit also applies a technique first done in SERVER-332 to enable
the ability to run SSLv3-based unit tests on newer JDK versions that
disable the use of SSLv3 by default.