(MAINT) DRY up connector instantiation
Prior to this commit, the logic for creating a Jetty ServerConnector
object was spread over two different places; one for HTTP, one for
HTTPS.  There was an increasing amount of logic duplicated in those
two places (specifically, request header size and so-linger timeout),
and there was going to be even more in order to implement the
idle timeout setting.  Therefore, in preparation for adding the
idle timeout setting, this commit refactors the connector creation
code so that the logic lives in a single helper function rather
than being duplicated.