(TK-151) Add support for idle timeout setting for Jetty
This commit adds a new setting, `idle-timeout-milliseconds`,
that can be used to control the idle timeout for a Jetty
connector.  If the setting is specified and Jetty detects a
socket that has had no activity for a period of time greater
than the value of this timeout setting, Jetty will close
the connection on the server side.