(MAINT) make http config tests easier to debug
This commit changes the implementation of all of the http config
tests to use `clojure.test`'s `(is (=` form, because that gives
you much more useful debugging output when a test fails.  (e.g.,
in Cursive you can see a really useful formatted diff of the
two inputs to the `=` test.)

Before this commit, the equality check was done inside of a function,
so the test output didn't give any useful output at all w/rt
what was actually different between the two maps.