(MAINT) move remaining constants from core to config
The last few constants that were left in the jetty-core namespace
seemed like they were going to be related to the work in TK-148
to unwind the defaults for max threads.  This commit just moves
those constants into the `config` namespace, next to the other
settings related to the max-threads stuff, so that they'll
be really obvious / visible when that work is taken on.