(TK-168) Fix bad logic in proxy client settings test
Prior to this commit, the logic I was using to try to validate
the settings for the HTTP client used by the proxy servlet
were not valid.  They were constructing an HttpClient instance
directly, when, in reality, the ProxyServlet code modifies some
settings on the client object that it instantiates after it
constructs it.

This commit changes these tests to build up a full proxy servlet
and then use that to construct the HttpClient instance that
we make the assertions against.