(MAINT) Make java.jmx a top-level dependency
This commit promotes org.clojure/java.jmx from a dev profile only to a
top-level dependency.  There is no real need for the production
tk-jetty9 jar to pull in the dependency at this point.  However, we
haven't found a way to have java.jmx be included as a direct dependency
of the 'testutils' jar such that users of the 'testutils' jar wouldn't
have to declare their own redundant dependency on org.clojure/java.jmx,
which is undesirable.