(TK-338) Catch and log error for TimeoutException during stop
When a TK app is sent a HUP signal to request a restart, Jetty will
attempt to finish off any active requests gracefully before shutting
down (and subsequently restarting).  However, there is a setting
called `stopTimeout` that determines how long the server will wait
to "gracefully" shut down, before it gets more aggressive in killing
off threads.

Prior to this commit, if we were in the midst of a HUP restart, and
this timeout was exceeded during the graceful shutdown attempt,
Jetty would throw a `TimeoutException` that was treated as fatal
by TK.  This commit catches that exception and logs an error message,
but allows the server to continue on and restart Jetty.

I've done a decent amount of manual testing on this (including an
overnight run that included a HUP every 10 minutes, and some
analysis of heap dumps taken at a few places during that run), and
it seems to behave well.

I need to add tests for this, but in the interest of time, I'm
currently planning to do that in a separate commit / PR.