(TK-343) handler wrapper and servlet filter for uri normalization
This commit adds 'middleware' functions for normalizing the value that
an HttpServletRequest.getRequestURI() call returns for downstream
handlers that receive the request.

For handlers that use a ContextHandler container - ring, websocket -
the middleware is a HandlerWrapper derivative,
`normalize-uri-handler`.

For handlers that use a derivative of a ServletContextHandler
container - static context, servlet, war - the middleware is a servlet
filter.