(TK-343) Removed jar-exclusions line from project.clj
This commit removes the jar-exclusions line from the project.clj.  This
causes the source file for one Java class currently in the project,
HttpServletRequestWithAlternateRequestUri, to be built into the jar.
We've been including Java source files inconsistently from project to
project anyway so it didn't seem critical to have this one explicitly
remove them.