(TK-343) Add exclamation point to add-filter-to-servlet-handler fn
This commit adds an exclamation point to the end of the name of the
add-normalized-uri-filter-to-servlet-handler! function in the
normalized-uri-helpers namespace, signifying that the function is only
called for the side effect of having a filter added to its handler
argument.

This commit also corrects the text in a testing string in one of the
uri encoding tests.