(TK-372) Reset MBeanContainer's map of bean names.
In the stop method of the WebserverService, clear the MBeanContainer
class's static map of bean name fragments, which it uses to ensure that
bean names are unique when a name is not specified when the bean is
registered. Without this change, this static map leaks these name
fragments unless JMX metrics are specifically disabled in the TK config.