(SERVER-1695) Expose a request-body-max-size webserver setting
This commit exposes a new webserver setting, request-body-max-size,
which can be used to configure the maximum HTTP Content-Length that
the Jetty webserver will accept for an incoming request.  If the
limit is exceeded, the request fails with a 413 error.  If the
setting is not present in configuration or the Content-Length
header is not present in the request, the request will be passed through
to underlying handlers for evaluation.