(TK-429) Allow gzip encoding when an access log is configured
Previously, when the `access-log-config` setting was specified in a
webserver config, gzip encoding was always disabled for responses (even
when `gzip-enabled` was set to true via configuration).  With the
changes in this PR, requests are routed through the GzipHandler whether
or not an access log is configured, allowing the `gzip-enabled` setting
to be properly honored regardless of the value of the
`access-log-config` setting.