(TK-369) Add LifeCycleImplementingRequestLogImpl
As of Jetty 9.3, logback's RequestLogImpl will no longer auto-start
because it no longer implements Jetty's LifeCycle interface (because
Jetty refactored the RequestLog interfact to no longer extend the
LifeCycle interface). In order for our request logger to auto-start, we
extend logback's implementation and have it implement Jetty's LifeCycle
interface (which it already does).