(TK-437) Reworked no graceful shutdown test to avoid hang
Previously, if the 'no graceful shutdown' test failed due to the server
not closing the socket when the server was stopped, this test could
hang.  This commit reworks the test steps to hopefully avoid hanging the
test suite in the event of this kind of failure.  There does still
appear to be an underlying bug in Jetty here but it has proven very
difficult to reliably reproduce.