(TK-149) Decrease sleep between crl read attempts
If the crl file were not readable or valid for some reason, we would
previously wait 1 second between read attempts. Whatever process was
updating the file would likely not need a full second to finish writing
to it. This commit updates the sleep to be 100 milliseconds, and
increases the number of read attempts to 25 to account for the change.