(TK-448) use non-recursive form of watch service watcher
This commit updates the add-watch-dir! call that we use to watch the crl
directory for changes to use the non-recursive option. We do so to address an
issue we were seeing where during the normal lifecycle of puppetserver run, it
was possible to create a directory that the watch service could not traverse,
causing the service to fail. With recursive false, we do not expect the service
to fail.

Signed-off-by: Moses Mendoza <moses@puppet.com>