(TK-470) Add logback-access pattern for MDC
This commit adds a logback-access pattern converter that provides access to
the SLF4J Mapped Diagnostic Context (MDC). The MDC is a thread-local map
that stores contextual data for easy access by loggers. The converter is
registered with the RequestLogHandler instances that log Jetty requests as a
handler for the %X and %mdc keywords.

This allows for access log configurations such as:

    <encoder>
      <pattern>... %X{jruby.borrow-time:-0}</pattern>
    </encoder>

Which would print any value of `jruby.borrow-time` stored in the MDC map or
0 if no value is stored.