(SERVER-2213) Remove minimum thread default test
In recent Jetty releases the minimum threads per connector has been
changed to vary depending on the size of the server's threadpool as well
as number of virtual cores. These tests became unreliable as a connector
may require as low as a minimum of two threads, but that will change
depending on the box. There may also be one or more worker threads
consumed. I'm not completely sure why, in testing on my machine, and
Travis a server with a single connector consumes 4 threads, one with
two connectors consumes 6, but in Jenkins a server with a single
connector consumes 5 threads. However, the default size of the server's
threadpool is 200, and you should NEVER run a production server with
only 4 or 5 threads so the this work seems to have diminished returns
for a contrived test case.