(maint) Remove so-linger-seconds option
As of Jetty v9.4.12 (specifically 4f1dd35[1]) Jetty no longer honors
the setSoLingerTime configuration value. It was discovered as the root
cause of a Windows bug report[2] that SO_LINGER has undefined behavior
for non-blocking IO and actual runtime behavior changes in Java 11[3].

Consequently, it is not only deemed ineffective, but potentially
harmful to allow users to set this.

When investigating why this option was originally added (TK-140[4])
there were no user cases given though anecdotally during that timeframe
there were many issues with zombie connections remaining on the master
because of, if we recall correctly, what were ultimately connection
handling issues on the Ruby client side.

1. https://github.com/eclipse/jetty.project/commit/4f1dd352d621667f272ac46ce8fccc8d4bb6c507
2. https://github.com/eclipse/jetty.project/issues/2468
3. http://mail.openjdk.java.net/pipermail/nio-dev/2018-June/005195.html
4. https://tickets.puppetlabs.com/browse/TK-140