(SERVER-2576) Loosen schema to allow so-linger-seconds key
Prior to 3.0 we allowed and claimed to honor setting the socket linger
setting. In 3.0 we removed this setting since it was removed in Jetty
after the maintainers found out that it had undefined behavior on their
target platforms.

Unfortunately, none of the projects that pull in tk-jetty9 wrap the
configuration parsing sufficiently to intercept this configuration and
provide a kinder upgrade path. Nor are those projects in a position to
break their config APIs in a release soon.

Consequently, we now allow the old config value in our schema, though
ignore it in further processing to provide a smoother upgrade path to
users.