(PE-26658) apply JVM security policy to FIPS
Add a FIPS profile, with a JVM security profile to limit the crypto
providers in an FIPS environment. Fixes several tests (and core
functions) to work in FIPS.

Removes support for SSLv3. It's been long enough since POODLE that we
should not support this is any form. You cannot enable it in FIPS mode
anyhow.