(SERVER-2800) Update Jetty to 9.4.28
This updates us to the newest Jetty verion.

This required two changes. One was test changes around CVE-2016-2785 and
the other involves our custom SslContextFactory.

For background on CVE-2016-2785, see
[TK-343](https://tickets.puppetlabs.com/browse/TK-343) about how certain
percent encodings could be used to create a directory traversal attack.

In [#4033](https://github.com/eclipse/jetty.project/issues/4033), the
Jetty maintainers, in response to a user who got an NPE when attempting
to serve files from his jar by percent encoding a jar path, loosened the
percent encoding rules. Quickly after merging those loosened validation
rules, it was reverted and the rules were additionally tightened up. As
part of that tightening, some of the dangerous percent encodings Jetty
used to let through (and we explicitly tested received an error from our
middleware) no longer are allowed. Despite removing the assertion that
we see the specific error message from our middleware we still check
that the directory attacks receive a 404 and the primary handler never
sees the request.

Previously, Jetty deprecated the SslContextFactory that we extend to
provide custom CRL reloading. It also moved the
endpointIdentificationAlgorithm setting to specialized inner classes,
either SslContextFactory.Server or SslContextFactory.Client depending on
what defaults are required. With the latest updates attempting to extend
the deprecated SslContextFactory raises and we must now extend the inner
class SslContextFactory.Server. As a side effect we now no longer need
to manually set the endpointIdentificationAlgorithm.