(maint) update clj-parent to 5.2.6
This updates clj-parent to 5.2.6 to include the new version of
bouncy-castle which has been renamed from `15on` to `18on` to
indicate that it is now only jdk8 and beyond compatible.

The unhandled exceptions are now consistent between fips
and non-fips which changed a couple of tests to not specialize.

Additionally the changelog was updated to prepare for a new
4.4.0 release.