(PDB-5645) Add Response to request map when :include-response set
If :include-response is true in an add-ring-handler call, then include
the pending request Response instance in the request map as
:puppetlabs.trapperkeeper.services.webserver.jetty9/response so that
the handler can interrogate/manipulate the response directly if
desired.

PuppetDB intends to use this to detect client disconnects via the
SocketChannel (via the Response getHttpChannel) in order to halt a
potentially expensive query sooner.